refactor(LottoPaymentService): service 패키지로 이동 및 이름 변경
